server/requirements.txt
fastapi==0.115.12
neo4j==5.26.0
pandas==2.2.3
pydantic==2.11.5
python-dotenv==1.1.0
rapidfuzz==3.13.0
Unidecode==1.3.8
uvicorn==0.34.2
